% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCI.R
\name{bootstrapCI}
\alias{bootstrapCI}
\title{Bootstrap Confidence Interval}
\usage{
bootstrapCI(data, alpha, n = 20)
}
\arguments{
\item{data}{a numeric dataset containing data for predictive modelling of a dataset}

\item{alpha}{the designated significance value}

\item{n}{the number of bootstraps (default is 20)}
}
\value{
a numeric matrix containing the bootstrap confidence interval
}
\description{
Bootstrap Confidence Interval
}
\examples{

data=read.csv("iris_csv.csv")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)

bootstrapCI(data,alpha=0.05,n=20)

}
