% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All functions ready for package building.R,
%   R/Predictor.R
\name{cost.glm}
\alias{cost.glm}
\title{Logistic Regression}
\usage{
cost.glm(theta, X)

cost.glm(theta, X)
}
\arguments{
\item{theta}{a numerical coefficient vector}

\item{X}{a matrix containing the training dataset independent variable parameters.}
}
\value{
a cost estimator of the data

a cost estimator of the data
}
\description{
Logistic Regression

Logistic Regression
}
