% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{glm_plot}
\alias{glm_plot}
\title{GLM Plot}
\usage{
glm_plot(ytest, xtest, i)
}
\arguments{
\item{ytest}{the numeric y values of the test dataset}

\item{xtest}{the numeric x values of the test dataset}

\item{i}{the column of x values to be graphed}
}
\value{
a line plot of the y values and one independent variable (indicated by `i`)
}
\description{
GLM Plot
}
\examples{
data=read.csv("iris_csv.csv")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)
n_features=ncol(data)
dt = sort(sample(nrow(data), nrow(data)*.7))
train<-data[dt,]
test<-data[-dt,]
x=as.matrix(train[,1:n_features-1])
y=train[,n_features]
xtest=test[,1:n_features-1]
ytest=test[,n_features]

glm_plot(ytest,xtest,2)

}
