% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log loss optim.R
\name{log_betas}
\alias{log_betas}
\title{Finding the Betas}
\usage{
log_betas(Data, Y, X)
}
\arguments{
\item{Data}{a \code{matrix} of data that we want to do a logistic regression on.}

\item{Y}{A \code{integer} representing the column of the data that you want to be the response variable}

\item{X}{A \code{vector} representing the column(s) of the data that you want to be the predictor variable(s).}
}
\value{
A \code{vector} giving the optimized beta values
}
\description{
This function will spit out the desired \code{beta} values for the
log loss function
}
\examples{
  #data obtained from https://stats.oarc.ucla.edu/r/dae/logit-regression/
  test_data <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
  test_resp <- 1
  test_pred <- 2:4
  log_betas(Data = test_data, Y = test_resp, X = test_pred)
}
\author{
Group 11
}
