% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log loss.R
\name{log_loss}
\alias{log_loss}
\title{Log Loss}
\usage{
log_loss(Beta, Obs, Resp, Preds)
}
\arguments{
\item{Beta}{A \code{vector} of numerics representing the beta values for the intercept and predictors for a logistic regression}

\item{Obs}{A \code{matrix} containing observations we want to
create a logistic model for.}

\item{Resp}{An \code{integer} that tells which column of the observation matrix you
want to be the response variable.}

\item{Preds}{A \code{vector} of integers corresponding to the columns for the desired predictor variables}
}
\value{
A \code{numeric} giving value of loss at \code{beta}
}
\description{
This function gives a \code{numeric} representing the loss given
a \code{vector} of observations and a \code{matrix} of coefficients
}
\author{
Group 11
}
