% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_LogReg.R
\name{logreg_optim}
\alias{logreg_optim}
\title{model fit}
\usage{
logreg_optim(x, y)
}
\arguments{
\item{x}{a numeric dataset containing explanatory variables}

\item{y}{a vector containing response variable}
}
\value{
a list of numeric coefficients
}
\description{
estimates logistic regression coefficients and intercept collectively called beta_hat.
}
