% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{sigmoid}
\alias{sigmoid}
\title{Fitted logistic curve}
\usage{
sigmoid(result)
}
\arguments{
\item{result}{a numeric data expression or dataset to be evaluated using a logistic curve.}
}
\value{
a vector of the predictors produced from `sigmoid` for each independent variable.
}
\description{
Calculates a predictive indicator of a dataset based on a logistic model.
}
\examples{
#' data=read.csv("iris_csv.csv")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)
n_features=ncol(data)
dt = sort(sample(nrow(data), nrow(data)*.7))
train<-data[dt,]
test<-data[-dt,]
x=as.matrix(train[,1:n_features-1])
y=train[,n_features]
xtest=test[,1:n_features-1]
ytest=test[,n_features]

init=weightInitialization(x,y)
x1 <- cbind(1, x)

g=sigmoid(x1\%*\%init)

}
