% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt function.R
\name{weightInitialization}
\alias{weightInitialization}
\title{Initialization of Optimization Values}
\usage{
weightInitialization(x, y)
}
\arguments{
\item{x}{a matrix containing the training dataset independent variable parameters.}

\item{y}{a numeric vector containing training dataset classification values.}
}
\value{
a matrix containing initial values for optimization.
}
\description{
`weightInitialization` returns initial values for optimization based on the least-squares formula.
}
\examples{
#' data=read.csv("iris_csv.csv")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)
n_features=ncol(data)
dt = sort(sample(nrow(data), nrow(data)*.7))
train<-data[dt,]
test<-data[-dt,]
x=as.matrix(train[,1:n_features-1])
y=train[,n_features]
xtest=test[,1:n_features-1]
ytest=test[,n_features]

init=weightInitialization(x,y)

}
