% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All functions ready for package building.R,
%   R/bootstrapCI.R
\name{bootstrapCI}
\alias{bootstrapCI}
\title{Title}
\usage{
bootstrapCI(data, alpha, n = 20)

bootstrapCI(data, alpha, n = 20)
}
\arguments{
\item{n}{}
}
\description{
Title

Title bootstrap CI
}
