% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_LogReg.R
\name{bootstrapCI}
\alias{bootstrapCI}
\title{Bootstrap Confidence Interval}
\usage{
bootstrapCI(x, y, alpha, n = 20)
}
\arguments{
\item{x}{a numeric dataset containing explanatory variables}

\item{y}{a vector containing response variable}

\item{alpha}{the designated significance level for the intervals}

\item{n}{the number of bootstraps (default is 20)}
}
\value{
a numeric matrix containing the bootstrap confidence interval
\item{beta}{estimates of the logistic regression model with different samples}
\item{CI}{confidence intervals for each variable and intercept}
}
\description{
calculates Bootstrap confidence intervals
}
\examples{

data=read.csv("iris_csv.csv")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)
dt = sort(sample(nrow(data), nrow(data)*.7))
train<-data[dt,]
x=as.matrix(train[,1:n_features-1])
y=train[,n_features]

bootstrapCI(x,y,alpha=0.05,n=20)

}
