% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{cost.glm}
\alias{cost.glm}
\title{Logistic Regression}
\usage{
cost.glm(theta, X)
}
\arguments{
\item{theta}{a numerical coefficient vector}

\item{X}{a matrix containing the training dataset independent variable parameters.}
}
\value{
a cost estimator of the data
}
\description{
Logistic Regression
}
\examples{
data=read.csv("iris_csv.csv")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)
n_features=ncol(data)
dt = sort(sample(nrow(data), nrow(data)*.7))
train<-data[dt,]
test<-data[-dt,]
x=as.matrix(train[,1:n_features-1])
y=train[,n_features]
xtest=test[,1:n_features-1]
ytest=test[,n_features]

init=weightInitialization(x,y)
x1 <- cbind(1, x)
predict=optim(par=init, fn = cost.glm, method='CG',
              X=x1)

cost.glm <- function(theta,X) {
 m <- nrow(X)
 g <- sigmoid(X\%*\%theta)
 (1/m)*sum((-y*log(g)) - ((1-y)*log(1-g)))
}

}
