% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_LogReg.R
\name{metricplot}
\alias{metricplot}
\title{metric Plot}
\usage{
metricplot(x, y, metric)
}
\arguments{
\item{x}{the numeric y values of the test dataset}

\item{y}{the predicted y values of the test dataset}

\item{metric}{is the name of mr=etric for which the plot is required, like Accuracy.}
}
\value{
any required metric plot against a range of cutoff values from 0.1-0.9
}
\description{
provides any required metric plot against a range of cutoff values.
}
\examples{
data=read.csv("iris_csv.csv")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)

n_features=ncol(data)
dt = sort(sample(nrow(data), nrow(data)*.7))
train<-data[dt,]
test<-data[-dt,]
x=as.matrix(train[,1:n_features-1])
y=train[,n_features]
xtest=test[,1:n_features-1]
ytest=test[,n_features]
metric_plot(x,y,metric="Accuracy")
}
