% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All functions ready for package building.R,
%   R/Predictor.R
\name{sigmoid}
\alias{sigmoid}
\title{Fitted logistic curve}
\usage{
sigmoid(result)

sigmoid(result)
}
\arguments{
\item{result}{a numeric data expression or dataset to be evaluated using a logistic curve.}
}
\value{
a vector of the predictors produced from `sigmoid` for each independent variable.

a vector of the predictors produced from `sigmoid` for each independent variable.
}
\description{
Calculates a predictive indicator of a dataset based on a logistic model.

Calculates a predictive indicator of a dataset based on a logistic model.
}
